Refactor: remove migration logic from localStorage to chrome APIs
