+++ src/background/devtools.ts
@@ -12,7 +12,6 @@ interface DevToolsStorage {
     set(key: string, value: string): void;
     remove(key: string): void;
     has(key: string): Promise<boolean>;
-    setDataIsMigratedForTesting(value: boolean): void;
 }
 
 declare const __DEBUG__: boolean;
@@ -21,72 +20,7 @@ class PersistentStorageWrapper implements DevToolsStorage {
     // Cache information within background context for future use without waiting.
     private cache: {[key: string]: string} = {};
 
-    // TODO(bershanskiy): remove migrated and migrateFromLocalStorage after migration end.
-    // Part 1 of 2.
-    private dataIsMigrated = false;
-
-    setDataIsMigratedForTesting(value: boolean) {
-        if (__DEBUG__) {
-            this.dataIsMigrated = value;
-        }
-    }
-
-    // This function moves DevTools data from loclStorage to chrome.storage.local.
-    // This function is run on every backgroun context invocation, but it has effect only on the
-    // first run.
-    private async migrateFromLocalStorage() {
-        // In MV3 world we can't access localStorage, so we can not migrate anything.
-        // Bail out and consider data migrated.
-        if (typeof localStorage === 'undefined') {
-            this.dataIsMigrated = true;
-            return;
-        }
-
-        return new Promise<void>((resolve) => {
-            chrome.storage.local.get([
-                DevTools.KEY_DYNAMIC,
-                DevTools.KEY_FILTER,
-                DevTools.KEY_STATIC
-            ], (data) => {
-                if (chrome.runtime.lastError) {
-                    console.error('DevTools failed to migrate data', chrome.runtime.lastError);
-                    resolve();
-                }
-                // If storage contains at least one relevant record, we consider data migrated.
-                if (data[DevTools.KEY_DYNAMIC] || data[DevTools.KEY_FILTER] || data[DevTools.KEY_STATIC]) {
-                    this.dataIsMigrated = true;
-                    this.cache = data;
-                    resolve();
-                    return;
-                }
-
-                this.cache = {
-                    [DevTools.KEY_DYNAMIC]: localStorage.getItem(DevTools.KEY_DYNAMIC),
-                    [DevTools.KEY_FILTER]: localStorage.getItem(DevTools.KEY_FILTER),
-                    [DevTools.KEY_STATIC]: localStorage.getItem(DevTools.KEY_STATIC),
-                };
-
-                chrome.storage.local.set(this.cache, () => {
-                    if (chrome.runtime.lastError) {
-                        console.error('DevTools failed to migrate data', chrome.runtime.lastError);
-                        resolve();
-                    }
-                    this.dataIsMigrated = true;
-                    // Clean up localStorage after migration
-                    localStorage.removeItem(DevTools.KEY_DYNAMIC);
-                    localStorage.removeItem(DevTools.KEY_FILTER);
-                    localStorage.removeItem(DevTools.KEY_STATIC);
-                    resolve();
-                });
-            });
-        });
-    }
-
     async get(key: string) {
-        if (!this.dataIsMigrated) {
-            await this.migrateFromLocalStorage();
-        }
-
         if (key in this.cache) {
             return this.cache[key];
         }
@@ -137,12 +71,6 @@ class PersistentStorageWrapper implements DevToolsStorage {
 }
 
 class LocalStorageWrapper implements DevToolsStorage {
-    setDataIsMigratedForTesting() {
-        if (__DEBUG__) {
-            logWarn('Unexpected call to setDataIsMigratedForTesting');
-        }
-    }
-
     async get(key: string) {
         try {
             return localStorage.getItem(key);
@@ -179,12 +107,6 @@ class LocalStorageWrapper implements DevToolsStorage {
 }
 
 class TempStorage implements DevToolsStorage {
-    setDataIsMigratedForTesting() {
-        if (__DEBUG__) {
-            logWarn('Unexpected call to setDataIsMigratedForTesting');
-        }
-    }
-
     map = new Map<string, string>();
 
     async get(key: string) {
@@ -224,15 +146,9 @@ export default class DevTools {
         this.onChange = onChange;
     }
 
-    // TODO(bershanskiy): make private again once PersistentStorageWrapper removes migration logic.
-    // Part 2 of 2.
-    static KEY_DYNAMIC = 'dev_dynamic_theme_fixes';
-    static KEY_FILTER = 'dev_inversion_fixes';
-    static KEY_STATIC = 'dev_static_themes';
-
-    setDataIsMigratedForTesting(value: boolean) {
-        this.store.setDataIsMigratedForTesting(value);
-    }
+    private static KEY_DYNAMIC = 'dev_dynamic_theme_fixes';
+    private static KEY_FILTER = 'dev_inversion_fixes';
+    private static KEY_STATIC = 'dev_static_themes';
 
     private async loadConfigOverrides() {
         const [
