+++ tests/browser/environment.js
@@ -225,7 +225,6 @@ class PuppeteerEnvironment extends JestNodeEnvironment.TestEnvironment {
                 getLocalStorage: async () => await sendToUIPage({type: 'getLocalStorage'}),
                 changeChromeStorage: async (region, data) => await sendToUIPage({type: 'changeChromeStorage', data: {region, data}}),
                 getChromeStorage: async (region, keys) => await sendToUIPage({type: 'getChromeStorage', data: {region, keys}}),
-                setDataIsMigratedForTesting: async (value) => await sendToUIPage({type: 'setDataIsMigratedForTesting', data: value}),
                 emulateMedia: async (name, value) => {
                     if (this.global.product === 'firefox') {
                         return;
