+++ tests/browser/globals.d.ts
@@ -23,7 +23,6 @@ declare global {
         getLocalStorage: () => Promise<{[key: string]: any}>;
         changeChromeStorage: (region: 'local' | 'sync', data: {[key: string]: any}) => Promise<void>;
         getChromeStorage: (region: 'local' | 'sync', keys: string[]) => Promise<{[key: string]: any}>;
-        setDataIsMigratedForTesting: (value: boolean) => Promise<void>;
         emulateMedia: (name: string, value: string) => Promise<void>;
     };
 }
